requirements.txt
django==4.2
djangorestframework==3.15.1
psycopg2-binary==2.9.10
Pillow==10.4.0
python-dotenv==1.0.1
drf-yasg==1.21.8
